refactor into main.cpp
