+++ lox/main.cpp
@@ -9,12 +9,35 @@
 
 #include "fmt/core.h"
 #include "fmt/format.h"
-#include "lox/run.hpp"
 #include "lyra/arg.hpp"
 #include "lyra/help.hpp"
 #include "lyra/lyra.hpp"
 #include "tl/expected.hpp"
 #include "tl/optional.hpp"
+
+void echo(std::istream &input) {
+  std::vector<std::string> input_lines{};
+
+  for (std::string line; std::getline(input, line);) {
+    input_lines.emplace_back(line);
+  }
+  fmt::print("{}\n\n", fmt::join(input_lines, "\n"));
+}
+
+void runFile(std::vector<std::string> const &filenames) {
+  for (auto const &filename : filenames) {
+    std::ifstream file(filename);
+    if (!file.is_open()) {
+      fmt::print(stderr, "Error opening file: {}\n", filename);
+      std::exit(1);
+    } else {
+      fmt::print("{}:\n", filename);
+      echo(file);
+    }
+  }
+}
+
+void runPrompt() { echo(std::cin); }
 
 void print_help() {
   fmt::print("lox - An interpreter for lox written in C++\n");
