--- lox/run.hpp
@@ -1,11 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-//
-// SPDX-FileCopyrightText: Copyright (c) assignUser
-#pragma once
-
-#include <string>
-#include <vector>
-
-void runPrompt();
-
-void runFile(std::vector<std::string> const &filenames);
